# be-nest/contracts/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/Session/CreateSession": {
      "post": {
        "operationId": "SessionController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSessionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "session"
        ]
      }
    },
...
    "/api/Session/UpdateStatusSession/{sessionId}": {
      "put": {
        "operationId": "SessionController_update",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSessionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "session"
        ]
      }
    },
    "/api/Session/DeleteSession/{sessionId}": {
      "delete": {
        "operationId": "SessionController_delete",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "session"
        ]
      }
    },
    "/api/Session/AutoFillSessionInDay": {
      "post": {
        "operationId": "SessionController_autoFill",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AutoFillSessionInDayDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "session"
        ]
      }
    },
...
    "/api/Notification": {
      "get": {
        "operationId": "NotificationController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "notification"
        ]
      },
      "post": {
        "operationId": "NotificationController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotificationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "notification"
        ]
      }
    },
...
    "/api/Notification/{id}": {
      "get": {
        "operationId": "NotificationController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "notification"
        ]
      },
      "put": {
        "operationId": "NotificationController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotificationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "notification"
        ]
      },
      "delete": {
        "operationId": "NotificationController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "notification"
        ]
      }
    },
    "/api/Notification/{id}/start-broadcast": {
      "post": {
        "operationId": "NotificationController_startBroadcast",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "notification"
        ]
      }
    },
    "/api/Notification/{id}/complete": {
      "post": {
        "operationId": "NotificationController_complete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "notification"
        ]
      }
    },
    "/api/Notification/{id}/cancel": {
      "post": {
        "operationId": "NotificationController_cancel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "notification"
        ]
      }
    },
    "/api/Notification/{id}/broadcast": {
      "post": {
        "operationId": "NotificationController_broadcast",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "notification"
        ]
      }
    },
...
  },
  "info": {
    "title": "Convocation Day API",
    "description": "NestJS compatibility API for the Convocation Day frontend",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "better-auth.session_token"
      }
    },
    "schemas": {
...
      "CreateSessionDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateSessionDto": {
        "type": "object",
        "properties": {}
      },
      "AutoFillSessionInDayDto": {
        "type": "object",
        "properties": {}
      },
...
      "NotificationDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}